alternate ISI/TO routine, optogenetic trials
added a sketch of how optogenetic trials will work, though they're not currently transmitted to serial which needs to happen.
Added dan's alternate vision of ISIs - both the exponentialish generator, TOs at all time, and TOs as more of a trial reset - although this only applies to ISI and optogenetics, NOT size and contrast. The full randomness of all but contrast may cause issues going forward.